Changed main logo font to Nakadai.
The license for Terminator NFI said $25 for commercial use and free for private use. JSFML is not commercial, but the license didn't make it clear at all whether people would have to pay for the font if they used JSFML in their commercial product. I switched to Nakadai now which is completely free. The M in "JSFML" looks sharper now. :)
Website and example scene have been updated.

Signed-off-by: pdinklag <pdinklag@googlemail.com>